silence conversion warnings
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/branches/RefactorRevGraph@14043 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/RevisionGraph/RemoveSimpleChanges.cpp
@@ -63,8 +63,9 @@ void CRemoveSimpleChanges::Apply (CVisibleGraph* graph, CVisibleGraphNode* node)
 {
     // "M", not a branch point, not the HEAD
 
-    if (   (node->GetClassification().Matches ( CNodeClassification::IS_MODIFIED
-                                              , CNodeClassification::MUST_BE_PRESERVED))
+    if (   (node->GetClassification().Matches 
+                ( CNodeClassification::IS_MODIFIED
+                , static_cast<DWORD>(CNodeClassification::MUST_BE_PRESERVED)))
         && (node->GetFirstTag() == NULL)
         && (node->GetFirstCopyTarget() == NULL))
     {
